Refactor polygon signed distance and start building SDF example
