+++ packages/examples/src/geometric-queries/signed-distance/signed-distance.domain
@@ -1,0 +1,2 @@
+type Polygon
+type Point
