+++ packages/examples/src/geometric-queries/signed-distance/signed-distance.substance
@@ -1,0 +1,301 @@
+Polygon P
+Point p0
+Point p1
+Point p2
+Point p3
+Point p4
+Point p5
+Point p6
+Point p7
+Point p8
+Point p9
+Point p10
+Point p11
+Point p12
+Point p13
+Point p14
+Point p15
+Point p16
+Point p17
+Point p18
+Point p19
+Point p20
+Point p21
+Point p22
+Point p23
+Point p24
+Point p25
+Point p26
+Point p27
+Point p28
+Point p29
+Point p30
+Point p31
+Point p32
+Point p33
+Point p34
+Point p35
+Point p36
+Point p37
+Point p38
+Point p39
+Point p40
+Point p41
+Point p42
+Point p43
+Point p44
+Point p45
+Point p46
+Point p47
+Point p48
+Point p49
+Point p50
+Point p51
+Point p52
+Point p53
+Point p54
+Point p55
+Point p56
+Point p57
+Point p58
+Point p59
+Point p60
+Point p61
+Point p62
+Point p63
+Point p64
+Point p65
+Point p66
+Point p67
+Point p68
+Point p69
+Point p70
+Point p71
+Point p72
+Point p73
+Point p74
+Point p75
+Point p76
+Point p77
+Point p78
+Point p79
+Point p80
+Point p81
+Point p82
+Point p83
+Point p84
+Point p85
+Point p86
+Point p87
+Point p88
+Point p89
+Point p90
+Point p91
+Point p92
+Point p93
+Point p94
+Point p95
+Point p96
+Point p97
+Point p98
+Point p99
+Point p100
+Point p101
+Point p102
+Point p103
+Point p104
+Point p105
+Point p106
+Point p107
+Point p108
+Point p109
+Point p110
+Point p111
+Point p112
+Point p113
+Point p114
+Point p115
+Point p116
+Point p117
+Point p118
+Point p119
+Point p120
+Point p121
+Point p122
+Point p123
+Point p124
+Point p125
+Point p126
+Point p127
+Point p128
+Point p129
+Point p130
+Point p131
+Point p132
+Point p133
+Point p134
+Point p135
+Point p136
+Point p137
+Point p138
+Point p139
+Point p140
+Point p141
+Point p142
+Point p143
+Point p144
+Point p145
+Point p146
+Point p147
+Point p148
+Point p149
+Point p150
+Point p151
+Point p152
+Point p153
+Point p154
+Point p155
+Point p156
+Point p157
+Point p158
+Point p159
+Point p160
+Point p161
+Point p162
+Point p163
+Point p164
+Point p165
+Point p166
+Point p167
+Point p168
+Point p169
+Point p170
+Point p171
+Point p172
+Point p173
+Point p174
+Point p175
+Point p176
+Point p177
+Point p178
+Point p179
+Point p180
+Point p181
+Point p182
+Point p183
+Point p184
+Point p185
+Point p186
+Point p187
+Point p188
+Point p189
+Point p190
+Point p191
+Point p192
+Point p193
+Point p194
+Point p195
+Point p196
+Point p197
+Point p198
+Point p199
+Point p200
+Point p201
+Point p202
+Point p203
+Point p204
+Point p205
+Point p206
+Point p207
+Point p208
+Point p209
+Point p210
+Point p211
+Point p212
+Point p213
+Point p214
+Point p215
+Point p216
+Point p217
+Point p218
+Point p219
+Point p220
+Point p221
+Point p222
+Point p223
+Point p224
+Point p225
+Point p226
+Point p227
+Point p228
+Point p229
+Point p230
+Point p231
+Point p232
+Point p233
+Point p234
+Point p235
+Point p236
+Point p237
+Point p238
+Point p239
+Point p240
+Point p241
+Point p242
+Point p243
+Point p244
+Point p245
+Point p246
+Point p247
+Point p248
+Point p249
+Point p250
+Point p251
+Point p252
+Point p253
+Point p254
+Point p255
+Point p256
+Point p257
+Point p258
+Point p259
+Point p260
+Point p261
+Point p262
+Point p263
+Point p264
+Point p265
+Point p266
+Point p267
+Point p268
+Point p269
+Point p270
+Point p271
+Point p272
+Point p273
+Point p274
+Point p275
+Point p276
+Point p277
+Point p278
+Point p279
+Point p280
+Point p281
+Point p282
+Point p283
+Point p284
+Point p285
+Point p286
+Point p287
+Point p288
+Point p289
+Point p290
+Point p291
+Point p292
+Point p293
+Point p294
+Point p295
+Point p296
+Point p297
+Point p298
+Point p299
